requirements.txt
streamlit>=1.22.0
crewai>=0.28.8
crewai-tools
chromadb>=0.3.0
sqlite-utils>=3.14.0
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
regex
langdetect
groq
google-generativeai
